+++ database/migrations/2024_02_02_130800_create_activities_table.php
@@ -11,23 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('activities', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('action_id')->nullable()->constrained('actions')->onDelete('cascade');
-            $table->foreignId('generic_model_id')->nullable()->constrained('generic_models')->onDelete('cascade');
-            $table->longText('properties')->nullable();
-            $table->string('unique_id')->nullable();
-            $table->string('generic_model_type')->nullable();
-            $table->string('description')->nullable();
-            $table->boolean('failed_job')->default(false);
-            $table->string('activity_type')->default('generic_model');
-            $table->index('unique_id');
-            $table->index('action_id');
-            $table->index('generic_model_id');
-            $table->index('generic_model_type');
-            $table->index('activity_type');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('activities')) {
+
+            Schema::create('activities', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('action_id')->nullable()->constrained('actions')->onDelete('cascade');
+                $table->foreignId('generic_model_id')->nullable()->constrained('generic_models')->onDelete('cascade');
+                $table->longText('properties')->nullable();
+                $table->string('unique_id')->nullable();
+                $table->string('generic_model_type')->nullable();
+                $table->string('description')->nullable();
+                $table->boolean('failed_job')->default(false);
+                $table->string('activity_type')->default('generic_model');
+                $table->index('unique_id');
+                $table->index('action_id');
+                $table->index('generic_model_id');
+                $table->index('generic_model_type');
+                $table->index('activity_type');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
